feat: furthered simplification efforts lol
