--- src/components/global-style/index.ts
@@ -1,7 +1,0 @@
-import { GlobalReset } from "./global-reset";
-import { GlobalStyle } from "./global-style";
-
-export {
-  GlobalReset,
-  GlobalStyle
-};
